getLastValueWithTimestamp: LastValueCache class was moved to a separate file
