+++ core/opendaq/signal/src/CMakeLists.txt
@@ -137,7 +137,7 @@ function(create_component_source_groups_${BASE_NAME})
     set(SDK_HEADERS_DIR "signal/include/opendaq")
     set(SDK_SRC_DIR "signal/src")
     
-    source_group("signal//signal" FILES 
+    source_group("signal//signal" FILES
         ${SDK_HEADERS_DIR}/signal.h
         ${SDK_HEADERS_DIR}/signal_impl.h
         ${SDK_HEADERS_DIR}/signal_container_impl.h
@@ -145,6 +145,7 @@ function(create_component_source_groups_${BASE_NAME})
         ${SDK_HEADERS_DIR}/signal_events.h
         ${SDK_HEADERS_DIR}/signal_private.h
         ${SDK_HEADERS_DIR}/signal_config.h
+        ${SDK_HEADERS_DIR}/last_value_cache.h
         ${SDK_SRC_DIR}/signal_impl.cpp
     )
     
@@ -354,6 +355,7 @@ set(SRC_PublicHeaders_Component
     packet_destruct_callback_impl.h
     packet_destruct_callback_factory.h
     signal_impl.h
+    last_value_cache.h
     input_port_impl.h
     connection_internal.h
     input_port_private.h
